Reworked the program to work in a menu
